requirements.txt
Pillow==8.3.1
python-slugify==5.0.2
Scrapy==2.5.0